Rename modules in prep for making pipable
