--- MacroSystem/MacroSystem.pyproj
@@ -1,38 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" DefaultTargets="Build">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>{1e0bc71d-a00e-45c9-a57d-a44a9e96d062}</ProjectGuid>
-    <ProjectHome />
-    <StartupFile />
-    <SearchPath />
-    <WorkingDirectory>.</WorkingDirectory>
-    <OutputPath>.</OutputPath>
-    <ProjectTypeGuids>{888888a0-9f3d-457c-b088-3a5042f75d52}</ProjectTypeGuids>
-    <LaunchProvider>Standard Python launcher</LaunchProvider>
-    <InterpreterId>Global|PythonCore|3.8-32</InterpreterId>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)' == 'Debug'" />
-  <PropertyGroup Condition="'$(Configuration)' == 'Release'" />
-  <PropertyGroup>
-    <VisualStudioVersion Condition=" '$(VisualStudioVersion)' == '' ">10.0</VisualStudioVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="core\natlinkmain.py" />
-    <Compile Include="core\redirect_output.py" />
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="core\msvcr100fix\README.txt" />
-    <Content Include="core\PYD\README.txt" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="core" />
-    <Folder Include="core\msvcr100fix" />
-    <Folder Include="core\PYD" />
-  </ItemGroup>
-  <ItemGroup>
-    <InterpreterReference Include="Global|PythonCore|3.8-32" />
-  </ItemGroup>
-  <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets" />
-</Project>
