+++ mypy.ini
@@ -1,5 +1,5 @@
 [mypy]
-files=MacroSystem/core/*.py,MacroSystem/core/*.pyi
+files=natlink/*.py,natlink/*.pyi
 follow_imports=error
 ignore_missing_imports=True
 disallow_untyped_defs=True
